+++ internal/serve/httphandler/stellar_toml_handler.go
@@ -25,6 +25,7 @@ type StellarTomlHandler struct {
 	NetworkPassphrase           string
 	Sep10SigningPublicKey       string
 	InstanceName                string
+	EnableAnchorPlatform        bool
 	BaseURL                     string
 }
 
@@ -49,21 +50,26 @@ func (s *StellarTomlHandler) buildGeneralInformation(ctx context.Context, req *h
 		accounts = fmt.Sprintf("[%q, %q]", perTenantDistributionAccount.Address, s.Sep10SigningPublicKey)
 	}
 
-	parsedBaseURL, err := url.Parse(s.BaseURL)
-	if err != nil {
-		log.Ctx(ctx).Warnf("Invalid environment BaseURL %s: %v", s.BaseURL, err)
-		parsedBaseURL = &url.URL{Scheme: "https"}
-	}
+	var webAuthEndpoint, transferServerSep0024 string
 
-	var webAuthEndpoint string
-	var transferServerSep0024 string
-	t, err := tenant.GetTenantFromContext(ctx)
-	if err != nil {
-		webAuthEndpoint = fmt.Sprintf("%s://%s/auth", parsedBaseURL.Scheme, req.Host)
-		transferServerSep0024 = fmt.Sprintf("%s://%s/sep24", parsedBaseURL.Scheme, req.Host)
+	if s.EnableAnchorPlatform {
+		webAuthEndpoint = s.AnchorPlatformBaseSepURL + "/auth"
+		transferServerSep0024 = s.AnchorPlatformBaseSepURL + "/sep24"
 	} else {
-		webAuthEndpoint = *t.BaseURL + "/auth"
-		transferServerSep0024 = *t.BaseURL + "/sep24"
+		parsedBaseURL, err := url.Parse(s.BaseURL)
+		if err != nil {
+			log.Ctx(ctx).Warnf("Invalid environment BaseURL %s: %v", s.BaseURL, err)
+			parsedBaseURL = &url.URL{Scheme: "https"}
+		}
+
+		t, err := tenant.GetTenantFromContext(ctx)
+		if err != nil {
+			webAuthEndpoint = fmt.Sprintf("%s://%s/auth", parsedBaseURL.Scheme, req.Host)
+			transferServerSep0024 = fmt.Sprintf("%s://%s/sep24", parsedBaseURL.Scheme, req.Host)
+		} else {
+			webAuthEndpoint = *t.BaseURL + "/auth"
+			transferServerSep0024 = *t.BaseURL + "/sep24"
+		}
 	}
 
 	return fmt.Sprintf(`