+++ CHANGELOG.md
@@ -24,6 +24,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 - Add `supported_assets` filter to `GET /wallets` endpoint [#734](https://github.com/stellar/stellar-disbursement-platform-backend/pull/734)
 - Add SEP10 /auth endpoints [#770](https://github.com/stellar/stellar-disbursement-platform-backend/pull/770)
 - Add OTP attempt tracking and validation with maximum attempts limit and expiration checks for receiver wallet registration [#773](https://github.com/stellar/stellar-disbursement-platform-backend/pull/773)
+- Add SEP24 /info endpoints [#773](https://github.com/stellar/stellar-disbursement-platform-backend/pull/773)
+- Add SEP24 interactive deposit endpoint [#802](https://github.com/stellar/stellar-disbursement-platform-backend/pull/802)
+- Add Anchor Platform enabled FF and remove docker support [#803](https://github.com/stellar/stellar-disbursement-platform-backend/pull/803)
+[#802](https://github.com/stellar/stellar-disbursement-platform-backend/pull/802)
+
 
 ### Changed